Create Step5 -
moment.js was the key for solving this for me. In the beginning I was hard coding the date calculation process. Using remainders and division signs to convert the interval into years, months, days, hours, minutes, and seconds. After a few days of research and continuous effort I realized that the way I chose to go was the wrong way. Then I researched date libraries for javascript and found moment. After learning how to import custom libraries to the IDE I was using it was only a matter of time before I came to the solution.